move rope params builder to op file
